requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
fsspec==2025.3.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.3
onnx==1.17.0
onnxruntime==1.21.0
opencv-python==4.11.0.86
packaging==24.2
paho-mqtt==2.1.0
pillow==11.1.0
protobuf==6.30.2
python-dotenv==1.1.0
requests==2.32.3
sympy==1.13.3
# torch==2.2.2
# torchvision==0.17.2
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
